Potential fix for code scanning alert no. 36: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type (size_t) to avoid overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done in the size_t type, which has a larger range and can accommodate larger values without overflow.

We will modify the line where the multiplication occurs, casting row to size_t before multiplying it by stride.

+++ libwebp-1.0.3/src/dsp/filters_neon.c
@@ -170,7 +170,7 @@ static WEBP_INLINE void DoGradientFilter_NEON(const uint8_t* in,
                                               int stride,
                                               int row, int num_rows,
                                               uint8_t* out) {
-  const size_t start_offset = row * stride;
+  const size_t start_offset = (size_t)row * stride;
   const int last_row = row + num_rows;
   SANITY_CHECK(in, out);
   in += start_offset;
